pdf-parser/requirements.txt
pdfplumber
python-doctr
python-doctr[torch]
docling